requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.7
django-bootstrap-icons==0.8.2
django-bootstrap-v5==1.0.11
django-jazzmin==2.6.0
django-localflavor==3.1
django-middleware-global-request==0.3.1
filelock==3.10.0
idna==3.4
pipenv==2023.3.20
platformdirs==3.1.1
psycopg2==2.9.5
python-stdnum==1.18
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
whitenoise==6.4.0
django-countries[pyuca]
ipython
django-isbn-field
Pillow
tqdm
django-solo==2.0.0
django-ajax-selects